Create README.md
WORLD CUP GROUP STAGE DRAW

The HTML document includes the following elements:

A title that displays "FIFA world cup".
A link to an external CSS file named "style.css".
A script tag that includes a link to the Chart.js library.
A header tag that includes a title with the text "FIFA GROUP STAGE DRAW".
An unordered list with the id "teams-list" that will be populated with the names of 16 teams participating in the world cup.
A header tag with the text "list of 16 Team".
A span tag for each team name with a class name that matches the country's name.
A div with the id "groups-list" that will display the drawn groups.
A button with the text "DRAW GROUP" that will trigger the "makeGroups" function to randomly assign teams to four groups and display them in the "groups-list" div.
A form that includes a button with the text "Give Points to compare".
A canvas element with the id "pie-chart" that will display the pie chart generated by the "makePieChart" function.

The JavaScript code includes the following functions:

"displayTeams" function that takes the "teams" array, clears the "teams-list" unordered list, and populates it with the team names.
"makeGroups" function that randomly shuffles the "teams" array and assigns four teams to each group. The function then displays the groups in the "groups-list" div.
"makePieChart" function that prompts the user to enter a group index (A-D) and then extracts the corresponding group from the "groups" array. The function then prompts the user to enter the points earned by each team in the selected group and generates a pie chart that compares the points of the teams. The pie chart is displayed in the "pie-chart" canvas element using the Chart.js library.
The code includes comments to explain the purpose of each function and variable.